added popularity indicator

+++ app/src/main/java/com/cerminnovations/moviesboard/data/remote/model/movie/Movie.kt
@@ -37,7 +37,7 @@ data class Movie(
     val backdropPath: String?,
 
     @Json(name = "popularity")
-    val popularity: Double,
+    val popularity: Float,
 
     @Json(name = "vote_count")
     val voteCount: Int,
